Issue #SB-0000 feat: Added new actors in conf file and deleted tasks from usermanagement actor
