+++ actors/user/src/main/java/org/sunbird/user/actors/UserManagementActor.java
@@ -29,7 +29,6 @@ import org.sunbird.common.models.util.LoggerEnum;
 import org.sunbird.common.models.util.ProjectLogger;
 import org.sunbird.common.models.util.ProjectUtil;
 import org.sunbird.common.models.util.ProjectUtil.EsType;
-import org.sunbird.common.models.util.ProjectUtil.Status;
 import org.sunbird.common.models.util.PropertiesCache;
 import org.sunbird.common.models.util.StringFormatter;
 import org.sunbird.common.models.util.datasecurity.DecryptionService;
@@ -39,7 +38,6 @@ import org.sunbird.common.request.ExecutionContext;
 import org.sunbird.common.request.Request;
 import org.sunbird.common.request.UserRequestValidator;
 import org.sunbird.common.responsecode.ResponseCode;
-import org.sunbird.common.responsecode.ResponseMessage;
 import org.sunbird.dto.SearchDTO;
 import org.sunbird.extension.user.UserExtension;
 import org.sunbird.extension.user.impl.UserProviderRegistryImpl;
@@ -64,13 +62,8 @@ import org.sunbird.telemetry.util.TelemetryUtil;
     "createUser",
     "updateUser",
     "getUserProfile",
-    "getRoles",
     "getUserDetailsByLoginId",
     "profileVisibility",
-    "unblockUser",
-    "blockUser",
-    "assignRoles",
-    "userCurrentLogin",
     "getMediaTypes"
   },
   asyncTasks = {}
@@ -108,18 +101,8 @@ public class UserManagementActor extends BaseActor {
       updateUser(request);
     } else if (operation.equalsIgnoreCase(ActorOperations.GET_PROFILE.getValue())) {
       getUserProfile(request);
-    } else if (operation.equalsIgnoreCase(ActorOperations.GET_ROLES.getValue())) {
-      getRoles();
     } else if (operation.equalsIgnoreCase(ActorOperations.GET_USER_DETAILS_BY_LOGINID.getValue())) {
       getUserDetailsByLoginId(request);
-    } else if (operation.equalsIgnoreCase(ActorOperations.BLOCK_USER.getValue())) {
-      blockUser(request);
-    } else if (operation.equalsIgnoreCase(ActorOperations.ASSIGN_ROLES.getValue())) {
-      assignRoles(request);
-    } else if (operation.equalsIgnoreCase(ActorOperations.UNBLOCK_USER.getValue())) {
-      unBlockUser(request);
-    } else if (operation.equalsIgnoreCase(ActorOperations.USER_CURRENT_LOGIN.getValue())) {
-      updateUserLoginTime(request);
     } else if (operation.equalsIgnoreCase(ActorOperations.GET_MEDIA_TYPES.getValue())) {
       getMediaTypes();
     } else if (operation.equalsIgnoreCase(ActorOperations.PROFILE_VISIBILITY.getValue())) {
@@ -328,22 +311,6 @@ public class UserManagementActor extends BaseActor {
         userId,
         dataMap);
     return true;
-  }
-
-  /**
-   * This method will update user current login time in keycloak
-   *
-   * @param actorMessage Request
-   */
-  private void updateUserLoginTime(Request actorMessage) {
-    String userId = (String) actorMessage.getRequest().get(JsonKey.USER_ID);
-    Response response = new Response();
-    response.put(JsonKey.RESPONSE, JsonKey.SUCCESS);
-    sender().tell(response, self());
-    if (Boolean.parseBoolean(PropertiesCache.getInstance().getProperty(JsonKey.IS_SSO_ENABLED))) {
-      boolean addedResponse = ssoManager.addUserLoginTime(userId);
-      ProjectLogger.log("user login time added response is ==" + addedResponse);
-    }
   }
 
   @SuppressWarnings("unchecked")
@@ -1831,65 +1798,6 @@ public class UserManagementActor extends BaseActor {
     reqMap.remove(JsonKey.ORGANISATION_ID);
   }
 
-  /** This method will provide the complete role structure.. */
-  @SuppressWarnings({"unchecked", "rawtypes"})
-  private void getRoles() {
-    Util.DbInfo roleDbInfo = Util.dbInfoMap.get(JsonKey.ROLE);
-    Util.DbInfo roleGroupDbInfo = Util.dbInfoMap.get(JsonKey.ROLE_GROUP);
-    Util.DbInfo urlActionDbInfo = Util.dbInfoMap.get(JsonKey.URL_ACTION);
-    Response mergeResponse = new Response();
-    List<Map<String, Object>> resposnemap = new ArrayList<>();
-    List<Map<String, Object>> list = null;
-    Response response =
-        cassandraOperation.getAllRecords(roleDbInfo.getKeySpace(), roleDbInfo.getTableName());
-    Response rolegroup =
-        cassandraOperation.getAllRecords(
-            roleGroupDbInfo.getKeySpace(), roleGroupDbInfo.getTableName());
-    Response urlAction =
-        cassandraOperation.getAllRecords(
-            urlActionDbInfo.getKeySpace(), urlActionDbInfo.getTableName());
-    List<Map<String, Object>> urlActionListMap =
-        (List<Map<String, Object>>) urlAction.getResult().get(JsonKey.RESPONSE);
-    List<Map<String, Object>> roleGroupMap =
-        (List<Map<String, Object>>) rolegroup.getResult().get(JsonKey.RESPONSE);
-    list = (List<Map<String, Object>>) response.getResult().get(JsonKey.RESPONSE);
-    if (list != null && !(list.isEmpty())) {
-      // This map will have all the master roles
-      for (Map<String, Object> map : list) {
-        Map<String, Object> roleResponseMap = new HashMap<>();
-        roleResponseMap.put(JsonKey.ID, map.get(JsonKey.ID));
-        roleResponseMap.put(JsonKey.NAME, map.get(JsonKey.NAME));
-        List<String> roleGroup = (List) map.get(JsonKey.ROLE_GROUP_ID);
-        List<Map<String, Object>> actionGroupListMap = new ArrayList<>();
-        roleResponseMap.put(JsonKey.ACTION_GROUPS, actionGroupListMap);
-        Map<String, Object> subRoleResponseMap = null;
-        for (String val : roleGroup) {
-          subRoleResponseMap = new HashMap<>();
-          Map<String, Object> subRoleMap = getSubRoleListMap(roleGroupMap, val);
-          List<String> subRole = (List) subRoleMap.get(JsonKey.URL_ACTION_ID);
-          List<Map<String, Object>> roleUrlResponList = new ArrayList<>();
-          subRoleResponseMap.put(JsonKey.ID, subRoleMap.get(JsonKey.ID));
-          subRoleResponseMap.put(JsonKey.NAME, subRoleMap.get(JsonKey.NAME));
-          for (String rolemap : subRole) {
-            roleUrlResponList.add(getRoleAction(urlActionListMap, rolemap));
-          }
-          if (subRoleResponseMap.containsKey(JsonKey.ACTIONS)) {
-            List<Map<String, Object>> listOfMap =
-                (List<Map<String, Object>>) subRoleResponseMap.get(JsonKey.ACTIONS);
-            listOfMap.addAll(roleUrlResponList);
-          } else {
-            subRoleResponseMap.put(JsonKey.ACTIONS, roleUrlResponList);
-          }
-          actionGroupListMap.add(subRoleResponseMap);
-        }
-
-        resposnemap.add(roleResponseMap);
-      }
-    }
-    mergeResponse.getResult().put(JsonKey.ROLES, resposnemap);
-    sender().tell(mergeResponse, self());
-  }
-
   /**
    * This method will find the action from role action mapping it will return action id, action name
    * and list of urls.
@@ -1943,199 +1851,6 @@ public class UserManagementActor extends BaseActor {
     return response;
   }
 
-  /**
-   * Method to block the user , it performs only soft delete from Cassandra , ES , Keycloak
-   *
-   * @param actorMessage
-   */
-  @SuppressWarnings("unchecked")
-  private void blockUser(Request actorMessage) {
-
-    ProjectLogger.log("Method call  " + "deleteUser");
-    Util.DbInfo usrDbInfo = Util.dbInfoMap.get(JsonKey.USER_DB);
-    Map<String, Object> userMap = (Map<String, Object>) actorMessage.getRequest().get(JsonKey.USER);
-    if (ProjectUtil.isNull(userMap.get(JsonKey.USER_ID))) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.invalidRequestData.getErrorCode(),
-              ResponseCode.invalidRequestData.getErrorMessage(),
-              ResponseCode.CLIENT_ERROR.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-    String userId = (String) userMap.get(JsonKey.USER_ID);
-    Map<String, Object> userDbRecord = Util.getUserbyUserId(userId);
-    if (null == userDbRecord) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.userNotFound.getErrorCode(),
-              ResponseCode.userNotFound.getErrorMessage(),
-              ResponseCode.RESOURCE_NOT_FOUND.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-
-    Map<String, Object> dbMap = new HashMap<>();
-    dbMap.put(JsonKey.IS_DELETED, true);
-    dbMap.put(JsonKey.STATUS, Status.INACTIVE.getValue());
-    dbMap.put(JsonKey.ID, userId);
-    dbMap.put(JsonKey.USER_ID, userId);
-    dbMap.put(JsonKey.UPDATED_DATE, ProjectUtil.getFormattedDate());
-    dbMap.put(JsonKey.UPDATED_BY, actorMessage.getRequest().get(JsonKey.REQUESTED_BY));
-
-    // deactivate from keycloak -- softdelete
-    if (isSSOEnabled) {
-      ssoManager.deactivateUser(dbMap);
-    }
-    // delete from registry
-    /*if (IS_REGISTRY_ENABLED) {
-      Map<String, Object> regMap = new HashMap<>();
-      regMap.put(JsonKey.REGISTRY_ID, userDbRecord.get(JsonKey.REGISTRY_ID));
-      UserExtension userExtension = new UserProviderRegistryImpl();
-      userExtension.delete(regMap);
-    }*/
-    // soft delete from cassandra--
-    Response response =
-        cassandraOperation.updateRecord(usrDbInfo.getKeySpace(), usrDbInfo.getTableName(), dbMap);
-    ProjectLogger.log("USER DELETED " + userId);
-    sender().tell(response, self());
-
-    // update record in elasticsearch ......
-    dbMap.remove(JsonKey.ID);
-    dbMap.remove(JsonKey.USER_ID);
-    ElasticSearchUtil.updateData(
-        ProjectUtil.EsIndex.sunbird.getIndexName(),
-        ProjectUtil.EsType.user.getTypeName(),
-        userId,
-        dbMap);
-    generateTeleEventForUser(null, userId, "blockUser");
-  }
-
-  /**
-   * This method will assign roles to users or user organizations.
-   *
-   * @param actorMessage
-   */
-  @SuppressWarnings("unchecked")
-  private void assignRoles(Request actorMessage) {
-    UserRequestValidator.validateAssignRole(actorMessage);
-    Map<String, Object> requestMap = actorMessage.getRequest();
-
-    if (null != requestMap.get(JsonKey.ROLES)
-        && !((List<String>) requestMap.get(JsonKey.ROLES)).isEmpty()) {
-      String msg = Util.validateRoles((List<String>) requestMap.get(JsonKey.ROLES));
-      if (!msg.equalsIgnoreCase(JsonKey.SUCCESS)) {
-        throw new ProjectCommonException(
-            ResponseCode.invalidRole.getErrorCode(),
-            ResponseCode.invalidRole.getErrorMessage(),
-            ResponseCode.CLIENT_ERROR.getResponseCode());
-      }
-    }
-
-    // object of telemetry event...
-    String userId = (String) requestMap.get(JsonKey.USER_ID);
-    String externalId = (String) requestMap.get(JsonKey.EXTERNAL_ID);
-    String provider = (String) requestMap.get(JsonKey.PROVIDER);
-    String organisationId = (String) requestMap.get(JsonKey.ORGANISATION_ID);
-    String hashTagId = null;
-    Map<String, Object> map = null;
-    // have a check if organisation id is provided then need to get hashtagId
-    if (StringUtils.isNotBlank(organisationId)) {
-      map =
-          ElasticSearchUtil.getDataByIdentifier(
-              ProjectUtil.EsIndex.sunbird.getIndexName(),
-              ProjectUtil.EsType.organisation.getTypeName(),
-              organisationId);
-      if (MapUtils.isNotEmpty(map)) {
-        hashTagId = (String) map.get(JsonKey.HASHTAGID);
-        requestMap.put(JsonKey.HASHTAGID, hashTagId);
-      }
-    } else {
-      SearchDTO searchDto = new SearchDTO();
-      Map<String, Object> filter = new HashMap<>();
-      filter.put(JsonKey.EXTERNAL_ID, externalId);
-      filter.put(JsonKey.PROVIDER, provider);
-      searchDto.getAdditionalProperties().put(JsonKey.FILTERS, filter);
-      Map<String, Object> esResponse =
-          ElasticSearchUtil.complexSearch(
-              searchDto,
-              ProjectUtil.EsIndex.sunbird.getIndexName(),
-              ProjectUtil.EsType.organisation.getTypeName());
-      List<Map<String, Object>> list = (List<Map<String, Object>>) esResponse.get(JsonKey.CONTENT);
-
-      if (!list.isEmpty()) {
-        map = list.get(0);
-        organisationId = (String) map.get(JsonKey.ID);
-        requestMap.put(JsonKey.ORGANISATION_ID, organisationId);
-        // get org hashTagId and keep inside request map.
-        hashTagId = (String) map.get(JsonKey.HASHTAGID);
-        requestMap.put(JsonKey.HASHTAGID, hashTagId);
-      }
-    }
-
-    // throw error if provided orgId or ExtenralId with Provider is not valid
-    if (MapUtils.isEmpty(map)) {
-      String errorMsg =
-          StringUtils.isNotEmpty(organisationId)
-              ? ProjectUtil.formatMessage(
-                  ResponseMessage.Message.INVALID_PARAMETER_VALUE,
-                  organisationId,
-                  JsonKey.ORGANISATION_ID)
-              : ProjectUtil.formatMessage(
-                  ResponseMessage.Message.INVALID_PARAMETER_VALUE,
-                  StringFormatter.joinByComma(externalId, provider),
-                  StringFormatter.joinByAnd(JsonKey.EXTERNAL_ID, JsonKey.PROVIDER));
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.invalidParameterValue.getErrorCode(),
-              errorMsg,
-              ResponseCode.CLIENT_ERROR.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-
-    // update userOrg role with requested roles.
-    Map<String, Object> userOrgDBMap = new HashMap<>();
-    userOrgDBMap.put(JsonKey.ORGANISATION_ID, organisationId);
-    userOrgDBMap.put(JsonKey.USER_ID, userId);
-    Util.DbInfo userOrgDb = Util.dbInfoMap.get(JsonKey.USER_ORG_DB);
-    Response response =
-        cassandraOperation.getRecordsByProperties(
-            userOrgDb.getKeySpace(), userOrgDb.getTableName(), userOrgDBMap);
-    List<Map<String, Object>> list = (List<Map<String, Object>>) response.get(JsonKey.RESPONSE);
-    if (list.isEmpty()) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.invalidUsrOrgData.getErrorCode(),
-              ResponseCode.invalidUsrOrgData.getErrorMessage(),
-              ResponseCode.CLIENT_ERROR.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-    // Add default role into Requested Roles if it is not provided and then update into DB
-    List<String> roles = (List<String>) requestMap.get(JsonKey.ROLES);
-    if (!roles.contains(ProjectUtil.UserRole.PUBLIC.name()))
-      roles.add(ProjectUtil.UserRole.PUBLIC.name());
-    userOrgDBMap.put(JsonKey.ROLES, roles);
-    userOrgDBMap.put(JsonKey.ID, list.get(0).get(JsonKey.ID));
-    if (StringUtils.isNotBlank(hashTagId)) {
-      userOrgDBMap.put(JsonKey.HASHTAGID, hashTagId);
-    }
-    userOrgDBMap.put(JsonKey.UPDATED_DATE, ProjectUtil.getFormattedDate());
-    userOrgDBMap.put(JsonKey.UPDATED_BY, requestMap.get(JsonKey.REQUESTED_BY));
-    userOrgDBMap.put(JsonKey.ROLES, roles);
-    response =
-        cassandraOperation.updateRecord(
-            userOrgDb.getKeySpace(), userOrgDb.getTableName(), userOrgDBMap);
-    sender().tell(response, self());
-    if (((String) response.get(JsonKey.RESPONSE)).equalsIgnoreCase(JsonKey.SUCCESS)) {
-      updateRoleToEs(userOrgDBMap, JsonKey.ORGANISATION, userId, organisationId);
-    } else {
-      ProjectLogger.log("no call for ES to save user");
-    }
-    generateTeleEventForUser(requestMap, userId, "userLevel");
-  }
-
   private void generateTeleEventForUser(
       Map<String, Object> requestMap, String userId, String objectType) {
     List<Map<String, Object>> correlatedObject = new ArrayList<>();
@@ -2182,90 +1897,6 @@ public class UserManagementActor extends BaseActor {
       ProjectLogger.log(
           "Exception Occurred during saving user to Es while joinUserOrganisation : ", ex);
     }
-  }
-
-  /**
-   * Method to un block the user
-   *
-   * @param actorMessage
-   */
-  @SuppressWarnings("unchecked")
-  private void unBlockUser(Request actorMessage) {
-
-    ProjectLogger.log("Method call  " + "UnblockeUser");
-    Util.getUserProfileConfig(systemSettingActorRef);
-    Util.DbInfo usrDbInfo = Util.dbInfoMap.get(JsonKey.USER_DB);
-    Map<String, Object> userMap = (Map<String, Object>) actorMessage.getRequest().get(JsonKey.USER);
-    if (ProjectUtil.isNull(userMap.get(JsonKey.USER_ID))) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.invalidRequestData.getErrorCode(),
-              ResponseCode.invalidRequestData.getErrorMessage(),
-              ResponseCode.CLIENT_ERROR.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-    String userId = (String) userMap.get(JsonKey.USER_ID);
-    Response resultFrUserId =
-        cassandraOperation.getRecordById(usrDbInfo.getKeySpace(), usrDbInfo.getTableName(), userId);
-    List<Map<String, Object>> dbResult =
-        (List<Map<String, Object>>) resultFrUserId.get(JsonKey.RESPONSE);
-    if (dbResult.isEmpty()) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.userNotFound.getErrorCode(),
-              ResponseCode.userNotFound.getErrorMessage(),
-              ResponseCode.RESOURCE_NOT_FOUND.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-    Map<String, Object> dbUser = dbResult.get(0);
-    if (dbUser.containsKey(JsonKey.IS_DELETED)
-        && isNotNull(dbUser.get(JsonKey.IS_DELETED))
-        && !((Boolean) dbUser.get(JsonKey.IS_DELETED))) {
-      ProjectCommonException exception =
-          new ProjectCommonException(
-              ResponseCode.userAlreadyActive.getErrorCode(),
-              ResponseCode.userAlreadyActive.getErrorMessage(),
-              ResponseCode.CLIENT_ERROR.getResponseCode());
-      sender().tell(exception, self());
-      return;
-    }
-
-    Map<String, Object> dbMap = new HashMap<>();
-    dbMap.put(JsonKey.IS_DELETED, false);
-    dbMap.put(JsonKey.STATUS, Status.ACTIVE.getValue());
-    dbMap.put(JsonKey.ID, userId);
-    dbMap.put(JsonKey.USER_ID, userId);
-    dbMap.put(JsonKey.UPDATED_DATE, ProjectUtil.getFormattedDate());
-    dbMap.put(JsonKey.UPDATED_BY, actorMessage.getRequest().get(JsonKey.REQUESTED_BY));
-
-    // Activate user from keycloak
-    if (isSSOEnabled) {
-      ssoManager.activateUser(dbMap);
-    }
-    // Activate user from cassandra-
-    Response response =
-        cassandraOperation.updateRecord(usrDbInfo.getKeySpace(), usrDbInfo.getTableName(), dbMap);
-    ProjectLogger.log("USER UNLOCKED " + userId);
-    sender().tell(response, self());
-
-    // update record in elasticsearch ......
-    if (((String) response.get(JsonKey.RESPONSE)).equalsIgnoreCase(JsonKey.SUCCESS)) {
-      ProjectLogger.log("UserManagementActor:unBlockUser : updating user data to ES.");
-      Request userRequest = new Request();
-      userRequest.setOperation(ActorOperations.UPDATE_USER_INFO_ELASTIC.getValue());
-      userRequest.getRequest().put(JsonKey.ID, userId);
-      try {
-        tellToAnother(userRequest);
-      } catch (Exception ex) {
-        ProjectLogger.log(
-            "UserManagementActor:unBlockUser : Exception occurred while unblocking user : ", ex);
-      }
-    } else {
-      ProjectLogger.log("UserManagementActor:unBlockUser : no call for ES to save user");
-    }
-    generateTeleEventForUser(null, userId, "unBlockUser");
   }
 
   /**
