+++ service/src/main/resources/application.conf
@@ -172,6 +172,24 @@ SunbirdMWSystem {
             nr-of-instances = 2
             dispatcher = rr-usr-dispatcher
           }
+        "/RequestRouter/*/UserStatusActor"
+          {
+            router = smallest-mailbox-pool
+            nr-of-instances = 2
+            dispatcher = rr-usr-dispatcher
+          }
+        "/RequestRouter/*/UserRoleActor"
+          {
+            router = smallest-mailbox-pool
+            nr-of-instances = 2
+            dispatcher = rr-usr-dispatcher
+          }
+        "/RequestRouter/*/UserLoginActor"
+          {
+            router = smallest-mailbox-pool
+            nr-of-instances = 2
+            dispatcher = rr-usr-dispatcher
+          }
         "/RequestRouter/*/ActorAuditLogServiceImpl"
           {
             router = smallest-mailbox-pool
